Hotel Reservation System
 I am happy to announce that I have completed my Internship project on the topic of Hotel Reservation System. In this system there are 3 types of room are available:- 
1) Single type
2)Double Type
3)Suite Type
You can choose from any type and also there is to avail food and AC room for the users. First, the starting price is shown to the users of the hotels and the available rooms for each type. Then after you book the available rooms get deducted from the rooms booked.